+++ Tests/test_file_tiff.py
@@ -70,6 +70,15 @@ class TestFileTiff:
             im.load()
             im.close()
 
+    def test_seek_after_close(self):
+        im = Image.open("Tests/images/multipage.tiff")
+        im.close()
+
+        with pytest.raises(ValueError):
+            im.n_frames
+        with pytest.raises(ValueError):
+            im.seek(1)
+
     def test_context_manager(self):
         with warnings.catch_warnings():
             with Image.open("Tests/images/multipage.tiff") as im:
@@ -705,6 +714,13 @@ class TestFileTiff:
 
         with Image.open(outfile) as reloaded:
             assert reloaded.info["icc_profile"] == icc_profile
+
+    def test_save_bmp_compression(self, tmp_path):
+        with Image.open("Tests/images/hopper.bmp") as im:
+            assert im.info["compression"] == 0
+
+            outfile = str(tmp_path / "temp.tif")
+            im.save(outfile)
 
     def test_discard_icc_profile(self, tmp_path):
         outfile = str(tmp_path / "temp.tif")